+++ src/main/java/nl/tno/oorti/ooencoder/omtcodec/RPRunsignedInteger32BECodec.java
@@ -20,7 +20,7 @@ class RPRunsignedInteger32BECodec extends HLAunsignedInteger32BECodec {
         .append("{")
         .append("\"type\" : ")
         .append('"')
-        .append(Short.class.getTypeName())
+        .append(Integer.class.getTypeName())
         .append('"')
         .append(",")
         .append("\"name\" : ")
