+++ src/main/java/nl/tno/oorti/ooencoder/omtcodec/RPRunsignedInteger64BECodec.java
@@ -20,7 +20,7 @@ class RPRunsignedInteger64BECodec extends HLAunsignedInteger64BECodec {
         .append("{")
         .append("\"type\" : ")
         .append('"')
-        .append(Short.class.getTypeName())
+        .append(Long.class.getTypeName())
         .append('"')
         .append(",")
         .append("\"name\" : ")
